+++ packaging/arch/omacontra.sh
@@ -1,2 +1,2 @@
 #!/bin/sh
-exec /usr/bin/python /usr/share/omacontra/main.py "$@"
+exec /usr/bin/python -E -s /usr/share/omacontra/main.py "$@"